// src/modules/rab/types/rab-csv-row.type.ts
export type RabCsvRow = {
  period: string;
  marcas: string;
  proprietarios: string | null;
  operadores: string | null;
  nrCertMatricula: string | null;
  nrSerie: string | null;
  cdTipo: string | null;
  dsModelo: string | null;
  nmFabricante: string | null;
  cdCls: string | null;
  nrPmd: string | null;
  cdTipoIcao: string | null;
  nrTripulacaoMin: string | null;
  nrPassageirosMax: string | null;
  nrAssentos: string | null;
  nrAnoFabricacao: string | null;
  dtValidadeCva: string | null;
  dtValidadeCa: string | null;
  dtCanc: string | null;
  dsMotivoCanc: string | null;
  cdInterdicao: string | null;
  dsGravame: string | null;
  dtMatricula: string | null;
  tpMotor: string | null;
  qtMotor: string | null;
  tpPouso: string | null;
  tpCa: string | null;
  cdPropositoCave: string | null;
  cfOperacional: string | null;
  dsCategoriaHomologacao: string | null;
  tpOperacao: string | null;
};